Migrated authentication from SQLite to PostgreSQL, added user roles and password reset support.
