+++ frontend/src/pages/Home.jsx
@@ -6,7 +6,7 @@ function Home() {
       <h1>Welcome to JedgeBot</h1>
       <p>Manage your trading strategies with ease.</p>
       <div className="buttons">
-        <Link to="/auth" className="btn">Enter</Link>
+        <Link to="/login" className="btn">Enter</Link>
       </div>
     </div>
   );
