+++ frontend/src/pages/LoginOld.jsx
@@ -1,0 +1,73 @@
+import { useState } from "react";
+import { useNavigate } from "react-router-dom"; // ✅ Import useNavigate
+const API_URL = import.meta.env.VITE_API_URL;
+
+const Login = () => {
+  const [username, setUsername] = useState("");
+  const [password, setPassword] = useState("");
+  const [error, setError] = useState("");
+  const navigate = useNavigate(); // ✅ Create navigation hook
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    setError("");
+    try {
+      const response = await fetch(`${API_URL}/auth/login`, {        method: "POST",
+        headers: {
+          "Content-Type": "application/x-www-form-urlencoded",
+        },
+        body: new URLSearchParams({ username, password }).toString(),
+      });
+
+      if (!response.ok) {
+        throw new Error("Login failed. Please check your credentials.");
+      }
+
+      const data = await response.json();
+      localStorage.setItem("access_token", data.access_token); // ✅ Store token
+      navigate("/dashboard"); // ✅ Redirect to dashboard
+      window.location.reload(); // ✅ Force UI update
+
+
+      navigate("/dashboard"); // ✅ Redirect user to dashboard
+    } catch (err) {
+      setError(err.message);
+    }
+  };
+
+  return (
+    <div className="flex items-center justify-center min-h-screen">
+      <div className="p-8 rounded-lg shadow-lg w-full max-w-md">
+        <h2 className="text-2xl font-bold text-center mb-6">Login</h2>
+        {error && <p className="text-red-500 text-center">{error}</p>}
+        <form onSubmit={handleSubmit} className="space-y-4">
+          <input
+            type="text"
+            placeholder="Username"
+            value={username}
+            onChange={(e) => setUsername(e.target.value)}
+            className="w-full px-4 py-2 border rounded-lg focus:outline-none"
+          />
+          <input
+            type="password"
+            placeholder="Password"
+            value={password}
+            onChange={(e) => setPassword(e.target.value)}
+            className="w-full px-4 py-2 border rounded-lg focus:outline-none"
+          />
+          <button
+            type="submit"
+            className="w-full bg-blue-600 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded-lg transition duration-200"
+          >
+            Login
+          </button>
+        </form>
+        <p className="text-center mt-4">
+          Don't have an account? <a href="/register" className="text-blue-400 hover:underline">Register</a>
+        </p>
+      </div>
+    </div>
+  );
+};
+
+export default Login;
