+++ pyproject.toml
@@ -26,6 +26,8 @@ pydantic = "^2.10.6"
 annotated-types = "^0.7.0"
 colorama = "^0.4.6"
 types-colorama = "^0.4.15.20240311"
+psycopg2-binary = "^2.9.10"
+types-psycopg2 = "^2.9.21.20250318"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.3.4"
